requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
pytest
python-dotenv
gunicorn